requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0

# Task Queue
celery[redis]==5.4.0
redis==5.0.0

# Storage
minio==7.2.0

# AI / ML
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Video processing
ffmpeg-python==0.2.0
scenedetect[opencv]>=0.6.0

# RunPod
runpod>=1.6.0

# Utilities
httpx>=0.27.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
tqdm>=4.66.0
mediapipe>=0.10.0
easy-dwpose>=1.0.1
